CDKv2 support
