+++ common.ts
@@ -1,0 +1,3 @@
+export class SesSmtpCredentialsProps {
+    readonly region: string;
+}
