+++ tsconfig.json
@@ -22,6 +22,9 @@
     "types": ["node", "jest"]
   },
   "include": [
-    "index.ts"
+    "index.ts",
+    "cdkv1.ts",
+    "cdkv2.ts",
+    "common.ts"
   ]
 }
